# time: improve performance of calendar algorithms and clarify its spec…