…ification

This implementation improves the performance of Go's time package using the recent algorithms developed by Neri and Schneider [1]. Their algorithms are faster than alternatives, my benchmarks confirm performance improvements in Go as well.

```
goos: linux
goarch: amd64
pkg: time
cpu: Intel(R) Xeon(R) CPU E5-2680 v2 @ 2.80GHz
                       | /tmp/master.txt |   /tmp/nerishcneider.txt
                       |     sec/op      |   sec/op    vs base
Now-8                        69.38n ± 0%   69.33n ± 0%  -0.07% (p=0.001 n=20)
NowUnixNano-8                70.35n ± 0%   70.08n ± 0%  -0.38% (p=0.000 n=20)
NowUnixMilli-8               71.47n ± 0%   71.26n ± 0%  -0.30% (p=0.000 n=20)
NowUnixMicro-8               71.32n ± 0%   71.21n ± 0%  -0.15% (p=0.000 n=20)
Format-8                     441.9n ± 0%   433.8n ± 0%  -1.83% (p=0.000 n=20)
FormatRFC3339-8              293.9n ± 0%   283.4n ± 0%  -3.57% (p=0.000 n=20)
FormatRFC3339Nano-8          297.5n ± 0%   286.9n ± 0%  -3.55% (p=0.000 n=20)
FormatNow-8                  242.4n ± 0%   239.7n ± 0%  -1.11% (p=0.000 n=20)
MarshalJSON-8                144.0n ± 0%   136.2n ± 0%  -5.42% (p=0.000 n=20)
MarshalText-8                142.0n ± 0%   136.2n ± 0%  -4.02% (p=0.000 n=20)
Parse-8                      241.2n ± 0%   240.9n ± 0%  -0.12% (p=0.002 n=20)
ParseRFC3339UTC-8            78.52n ± 0%   73.12n ± 0%  -6.87% (p=0.000 n=20)
ParseRFC3339UTCBytes-8       86.60n ± 0%   81.05n ± 0%  -6.42% (p=0.000 n=20)
ParseRFC3339TZ-8             291.5n ± 0%   283.1n ± 0%  -2.90% (p=0.000 n=20)
ParseRFC3339TZBytes-8        343.2n ± 0%   333.2n ± 0%  -2.93% (p=0.000 n=20)
ParseDuration-8              125.6n ± 0%   125.8n ± 0%  +0.16% (p=0.002 n=20)
Hour-8                       6.458n ± 0%   6.457n ± 0%       ~ (p=0.130 n=20)
Second-8                     6.460n ± 0%   6.456n ± 0%  -0.05% (p=0.002 n=20)
Year-8                       18.68n ± 0%   14.10n ± 0% -24.52% (p=0.000 n=20)
Day-8                        24.01n ± 0%   17.14n ± 0% -28.61% (p=0.000 n=20)
ISOWeek-8                    27.55n ± 0%   24.46n ± 0% -11.22% (p=0.000 n=20)
GoString-8                   157.9n ± 0%   152.9n ± 0%  -3.13% (p=0.000 n=20)
UnmarshalText-8              290.1n ± 0%   283.4n ± 0%  -2.31% (p=0.000 n=20)
Geomean                      93.84n ± 0%   89.08n ± 0%  -5.07%
```

To achieve better performance we should change the supported range of dates: instead of supporting years taking any int value, they should be restricted to a much smaller range which still spans tens of thousands of years.

[1] Neri, C, Schneider, L. Euclidean affine functions and their application to calendar algorithms. Softw Pract Exper. 2023; 53(4): 937–970. doi:10.1002/spe.3172 https://onlinelibrary.wiley.com/doi/full/10.1002/spe.3172

For: golang/go#63844

This PR will be imported into Gerrit with the title and first
comment (this text) used to generate the subject and body of
the Gerrit change.

**Please ensure you adhere to every item in this list.**

More info can be found at https://github.com/golang/go/wiki/CommitMessage

+ The PR title is formatted as follows: `net/http: frob the quux before blarfing`
  + The package name goes before the colon
  + The part after the colon uses the verb tense + phrase that completes the blank in,
    "This change modifies Go to ___________"
  + Lowercase verb after the colon
  + No trailing period
  + Keep the title as short as possible. ideally under 76 characters or shorter
+ No Markdown
+ The first PR comment (this one) is wrapped at 76 characters, unless it's
  really needed (ASCII art, table, or long link)
+ If there is a corresponding issue, add either `Fixes #1234` or `Updates #1234`
  (the latter if this is not a complete fix) to this comment
+ If referring to a repo other than `golang/go` you can use the
  `owner/repo#issue_number` syntax: `Fixes golang/tools#1234`
+ We do not use Signed-off-by lines in Go. Please don't add them.
  Our Gerrit server & GitHub bots enforce CLA compliance instead.
+ Delete these instructions once you have read and applied them